ai-job-assistant/requirements.txt
fastapi
uvicorn
langgraph
langchain-core
tavily-python
ollama
